Update the requirements
I'm not entirely sure what this will actually work down to, but the environment this was built in was 5.3 and my server is 5.6, so it should be good for all of the versions in that range without issue.  I'm keeping an eye on 7, but this should be good on 7 assuming nothing major changes at the last second.